Kokkos Versions: fix illegal initialization of Kokkos Views
The View handles are allocated as part of a struct using malloc in a
C function. This leaves the views uninitialized. Later new Views are
assigned to those uninitialized ones. This is illegal behaviour.
The Views constructor must be called before an assignment operator can
be used. For a while this would just lead to undefined behavior in
certain corner cases, with the latest Kokkos this will lead typically
to segmentation faults. The usage scenario can still be supported by
using a placement new operator for the first assignment operations.